feat: refactor data attribute on nodes
In its current form, adding additional data means you have to access it like `node.value.data.someValue` which is very inconvenient. In this update we rename `value` to `data` and merge top-level data values with user defined values. This allows data to be accessed like `node.data.someValue` which is a much cleaner interface.